put everything in src, refactor lambda handler and use environment variable for connections string
