+++ src/services/inventory_service.py
